+++ zest/releaser/utils.py
@@ -32,6 +32,7 @@ WRONG_IN_VERSION = ['svn', 'dev', '(']
 MUST_CLOSE_FDS = not sys.platform.startswith('win')
 
 AUTO_RESPONSE = False
+PACKAGE_ROOT = os.curdir
 VERBOSE = False
 INPUT_ENCODING = 'UTF-8'
 if getattr(sys.stdin, 'encoding', None):
@@ -133,7 +134,12 @@ def cleanup_version(version):
 def parse_options():
     global AUTO_RESPONSE
     global VERBOSE
+    global PACKAGE_ROOT
     parser = ArgumentParser()
+    parser.add_argument(
+        "--package-root",
+        default=PACKAGE_ROOT,
+        help="Directory containing setup.py")
     parser.add_argument(
         "--no-input",
         action="store_true",
@@ -150,6 +156,7 @@ def parse_options():
     options = parser.parse_args()
     AUTO_RESPONSE = options.auto_response
     VERBOSE = options.verbose
+    PACKAGE_ROOT = options.package_root
 
 
 # Hack for testing, see get_input()
@@ -371,7 +378,9 @@ def setup_py(rest_of_cmdline):
             if unsafe in rest_of_cmdline:
                 executable = 'echo MOCK'
 
-    return '%s setup.py %s' % (executable, rest_of_cmdline)
+    setup_py = os.path.join(PACKAGE_ROOT, 'setup.py')
+
+    return ' '.join([executable, setup_py, rest_of_cmdline])
 
 
 def twine_command(rest_of_cmdline):