+++ CHANGES.rst
@@ -22,6 +22,8 @@ Changelog for zest.releaser
   wheel.  Issue #108.
   [maurits]
 
+- Use the AST and tokenizer when parsing and writing Python code files.
+  [mitchellrj]
 
 5.0 (2015-06-05)
 ----------------
@@ -32,7 +34,6 @@ Changelog for zest.releaser
 - Use the same `readme` library that PyPI uses to parse long
   descriptions when we test and render them.
   [mitchellrj]
-
 
 4.0 (2015-05-21)
 ----------------