+++ CHANGES.rst
@@ -16,7 +16,8 @@ Changelog for zest.releaser
   wheel.  Issue #108.
   [maurits]
 
-- Add support for projects with setup.py located in subdirectories.
+- Add a `--package-root` option to support projects where `setup.py` is not
+  located at the top level of the repository.
 
 
 5.0 (2015-06-05)